rename getBookName >> getBookMeta
